deleted manageUserAccount folder from pages folder, deleted common folder and moved userAPI to API folder, fixed path issues due to the userAPI move, fixed manage user error that didnt display user info when pressing edit profile, some css for manage user account page, created button folder for button component.
