--- src/Pages/ManageUserAccount/EditPicModal.css
@@ -1,44 +1,0 @@
-.editPicMod {
-  background-color: rgb(0, 0, 0);
-  background-color: rgba(0, 0, 0, 0.4);
-  display: none;
-  height: 100%;
-  left: 0;
-  overflow: auto;
-  position: fixed;
-  top: 0;
-  width: 100%;
-  z-index: 1;
-}
-
-.picModContent {
-  background-color: #fefefe;
-  border: 1px solid #888;
-  margin: 15% auto;
-  padding: 25px;
-  width: 50%;
-}
-.close {
-  color: #aaa;
-  float: right;
-  font-size: 28px;
-  font-weight: bold;
-}
-
-.close:hover,
-.close:focus {
-  color: black;
-  cursor: pointer;
-  text-decoration: none;
-}
-
-#myBtn {
-  background-color: whitesmoke;
-  border: none;
-  border-radius: 30px;
-  color: rgb(53, 53, 53);
-  padding: 10px;
-  position: absolute;
-  right: 150px;
-  top: 250px;
-}
