+++ src/ReactComponents/Button/Button.css
@@ -1,0 +1,22 @@
+.button {
+  background-color: #f2f2f2;
+  border: 1px solid #5271ff;
+  border-radius: 7px;
+  color: #5271ff;
+  padding: 10px 5px;
+  width: 100px;
+}
+
+.button:hover {
+  background-color: #385dff;
+  color: #f2f2f2;
+}
+
+.button:active {
+  background-color: #5271ff;
+  border: none;
+  border-radius: 1.5rem;
+  color: #f2f2f2;
+  padding: 10px 10px;
+  width: 100px;
+}
